requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.38.0
python-multipart>=0.0.12
python-socketio>=5.11.0
python-engineio>=4.9.0

# Async Support
aiofiles>=24.1.0
aiohttp>=3.10.0
anyio>=4.4.0
httpx>=0.27.0
httpcore>=1.0.5

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.31
alembic>=1.13.0

# Vector Database
qdrant-client>=1.10.0
chromadb>=0.5.0
redis>=5.0.1

# Sentence Transformers (for embeddings)
sentence-transformers>=3.0.0

# AI/ML Core
openai>=1.40.0
langchain>=0.2.12
langchain-community>=0.2.10
langchain-core>=0.2.27
langchain-openai>=0.1.20
langchain-text-splitters>=0.2.2
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
langgraph-sdk>=0.1.20
langsmith>=0.1.93

# CrewAI and Agents
crewai>=0.55.0
instructor>=1.3.0
litellm>=1.43.0
dspy-ai>=2.4.12

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.11.0
pdfminer.six>=20231228
pypdfium2>=4.30.0

# Web Scraping/Crawling
Crawl4AI>=0.3.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
playwright>=1.45.0
fake-useragent>=1.5.0
fake-http-header>=0.3.5
cssselect>=1.2.0

# Computer Vision & ML
opencv-python-headless>=4.10.0.84
deepface>=0.0.91
mediapipe>=0.10.14
tf-keras>=2.17.0
ultralytics>=8.2.0
ultralytics-thop>=2.0.0
torch>=2.3.0
torchvision>=0.18.0
onnxruntime>=1.18.0
pillow>=10.4.0
scikit-learn>=1.5.1

# NLP
nltk>=3.8.1
tiktoken>=0.7.0
tokenizers>=0.19.0
regex>=2024.5.15

# Data Processing
numpy<2
pandas>=2.2.2
scipy>=1.14.0
polars>=0.20.0
pyarrow>=16.0.0
openpyxl>=3.1.5

# Hugging Face
huggingface_hub>=0.24.0
datasets>=2.20.0
transformers>=4.43.0
safetensors>=0.4.3

# Audio Processing
librosa>=0.10.2
torchaudio>=2.3.0
audioread>=3.0.1
soundfile>=0.12.1
soxr>=0.3.7

# AutoGen
pyautogen>=0.2.30
flaml>=2.2.0
docker>=7.1.0
termcolor>=2.4.0

# Performance
numba>=0.60.0
llvmlite>=0.43.0
pooch>=1.8.2
msgpack>=1.0.8
lazy_loader>=0.4
decorator>=5.1.1

# Authentication & Security
bcrypt>=4.2.0
PyJWT>=2.9.0
cryptography>=43.0.0
pyOpenSSL>=24.2.1

# Configuration & Environment
python-dotenv>=1.0.1
python-decouple>=3.8
pydantic>=2.8.0
pydantic-settings>=2.4.0
pydantic_core>=2.20.0

# HTTP & API
requests>=2.32.0
urllib3>=2.2.0
websockets>=12.0
websocket-client>=1.8.0
sse-starlette>=2.1.0
httpx-sse>=0.4.0

# Kubernetes
kubernetes>=30.1.0
google-auth>=2.32.0

# CLI & Utilities
click>=8.1.7
rich>=13.7.0
typer>=0.12.0
tqdm>=4.66.0
humanize>=4.10.0
colorlog>=6.8.0
coloredlogs>=15.0.1

# JSON Processing
orjson>=3.10.0
ujson>=5.10.0
json5>=0.9.25
json_repair>=0.25.0
jsonschema>=4.23.0

# Serialization
marshmallow>=3.21.0
dataclasses-json>=0.6.7

# Caching
diskcache>=5.6.3
cachetools>=5.4.0

# Telemetry & Monitoring
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-semantic-conventions>=0.46b0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
opentelemetry-exporter-otlp-proto-common>=1.25.0
opentelemetry-proto>=1.25.0
posthog>=3.5.0

# Templating
Jinja2>=3.1.4
MarkupSafe>=2.1.5
Mako>=1.3.5

# Search & Ranking
rank-bm25>=0.2.2
mmh3>=4.1.0

# Geometry & Visualization
matplotlib>=3.9.0
shapely>=2.0.5
trimesh>=4.4.0
alphashape>=1.3.1
rtree>=1.3.0

# MCP
mcp>=1.0.0

# gRPC
grpcio>=1.65.0
protobuf>=4.25.0

# Optuna (ML hyperparameter tuning)
optuna>=3.6.0

# Other Dependencies
PyYAML>=6.0.1
certifi>=2024.7.4
charset-normalizer>=3.3.0
idna>=3.7
six>=1.16.0
attrs>=24.2.0
filelock>=3.15.0
fsspec>=2024.6.0
networkx>=3.3
sympy>=1.13.0
packaging>=24.1
typing_extensions>=4.12.0
platformdirs>=4.2.0
tenacity>=8.5.0
sniffio>=1.3.1
backoff>=2.2.1
greenlet>=3.0.0
joblib>=1.4.0
threadpoolctl>=3.5.0
xxhash>=3.4.0
psutil>=6.0.0
deprecated>=1.2.14
wrapt>=1.16.0
